Regenerate the Hotspot numbers
The pljava.vmoptions had -Djava.home pointing into the /var/tmp
ramfs, but the compiled-in pljava.libjvm_location still pointed
to the shared object on disk.

Also, PL/Java-loaded 'saxon' jar need not be on the PL/Java classpath
during the Hotspot tests, when the Saxon jar is on the system classpath.